fix : 사진만 수정할 때 Null pointer exception 문제 해결

+++ footprint/src/main/java/com/umc/footprint/src/footprints/FootprintService.java
@@ -45,7 +45,9 @@ public class FootprintService {
             int userIdx = footprintDao.findUserIdx(footprintIdx);
 
             // 1. 본문 수정
-            footprintDao.modifyWrite(patchFootprintReq, footprintIdx);
+            if(patchFootprintReq.getWrite() != null) {
+                footprintDao.modifyWrite(patchFootprintReq, footprintIdx);
+            }
 
             // 2. 사진 수정
             // DB에 저장되어 있는 파일 리스트
@@ -76,18 +78,20 @@ public class FootprintService {
             // 전달되어온 태그 리스트
             List<String> tags = patchFootprintReq.getTagList();
 
-            if(dbTagList.isEmpty()) { // 발자국에 저장된 태그가 존재하지 않음
-                if(!tags.isEmpty()) { // 전달된 태그가 하나라도 존재
-                    footprintDao.addTag(tags, userIdx, footprintIdx);
+            if(patchFootprintReq.getTagList() != null) {
+                if(dbTagList.isEmpty()) { // 발자국에 저장된 태그가 존재하지 않음
+                    if(!tags.isEmpty()) { // 전달된 태그가 하나라도 존재
+                        footprintDao.addTag(tags, userIdx, footprintIdx);
+                    }
                 }
-            }
-            else { // 발자국에 저장된 기존 사진들이 존재
-                if(tags.isEmpty()) { // 전달된 파일이 없음 > 사진을 지우고 싶다는 의미 > 기존 사진 삭제만 진행
-                    footprintDao.deleteHashtags(footprintIdx); // 기존 해시태그들 테이블에서 삭제
-                }
-                else {
-                    footprintDao.deleteHashtags(footprintIdx);
-                    footprintDao.addTag(tags, userIdx, footprintIdx); // 새로운 사진들 업로드
+                else { // 발자국에 저장된 기존 사진들이 존재
+                    if(tags.isEmpty()) { // 전달된 파일이 없음 > 사진을 지우고 싶다는 의미 > 기존 사진 삭제만 진행
+                        footprintDao.deleteHashtags(footprintIdx); // 기존 해시태그들 테이블에서 삭제
+                    }
+                    else {
+                        footprintDao.deleteHashtags(footprintIdx);
+                        footprintDao.addTag(tags, userIdx, footprintIdx); // 새로운 사진들 업로드
+                    }
                 }
             }
